Replaced all C/C++ mentions to just C++
Removed ALL references to C/C++ and changed them to just say "C++". The
reason for this is because a C decompiler does not exist anywhere in
this source tree, so saying "C/C++" is not just misleading and
unhelpful, it's factually incorrect.

C and C++ are not the same language, are not similar, and are not even
compatible, and I could give countless examples to that end.